# docs/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/profiles/{steamId}/snapshots": {
      "get": {
        "description": "Returns stored aggregate snapshots for a Steam profile.",
        "operationId": "listProfileSnapshots",
        "parameters": [
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "steamId",
            "required": true,
            "in": "path",
            "description": "Steam 64-bit profile id",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileSnapshotsResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Profile was not found"
          }
        },
        "summary": "List profile snapshots",
        "tags": [
          "snapshots"
        ]
      },
      "post": {
        "description": "Creates a snapshot from current stored profile progress. Requires an authenticated owner of the claimed Steam profile, or an admin/moderator account.",
        "operationId": "createProfileSnapshot",
        "parameters": [
          {
            "name": "steamId",
            "required": true,
            "in": "path",
            "description": "Steam 64-bit profile id",
            "schema": {}
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileSnapshotResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "Authenticated user does not own this Steam profile."
          },
          "404": {
            "description": "Profile was not found"
          }
        },
        "summary": "Create a manual profile snapshot",
        "tags": [
          "snapshots"
        ]
      }
    },
...
  },
  "info": {
    "title": "Steam Achievement Tracker API",
    "description": "API for syncing and reading Steam achievement intelligence data",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Docker backend"
    }
  ],
  "components": {
    "schemas": {
...
      "ProfileSnapshotResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "totalGames": {
            "type": "number"
          },
          "completedGames": {
            "type": "number"
          },
          "totalAchievements": {
            "type": "number"
          },
          "unlockedAchievements": {
            "type": "number"
          },
          "remainingAchievements": {
            "type": "number"
          },
          "averageCompletionPercentage": {
            "type": "number"
          },
          "totalPlaytimeMinutes": {
            "type": "number"
          },
          "rarestUnlockedGlobalPercentage": {
            "type": "number",
            "nullable": true
          },
          "snapshotReason": {
            "type": "string",
            "enum": [
              "manual",
              "sync_completed",
              "scheduled"
            ]
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "totalGames",
          "completedGames",
          "totalAchievements",
          "unlockedAchievements",
          "remainingAchievements",
          "averageCompletionPercentage",
          "totalPlaytimeMinutes",
          "rarestUnlockedGlobalPercentage",
          "snapshotReason",
          "createdAt"
        ]
      },
      "ProfileSnapshotsResponseDto": {
        "type": "object",
        "properties": {
          "steamId": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProfileSnapshotResponseDto"
            }
          },
          "total": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "offset": {
            "type": "number"
          }
        },
        "required": [
          "steamId",
          "items",
          "total",
          "limit",
          "offset"
        ]
      },
...
    }
  }
}